Devided block into block-components, added custom components feature
