+++ examples/minimal/components/NotionPage.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react'
 import Head from 'next/head'
 
+import { NotionRenderer } from 'areyes-react-notion-x'
 import { ExtendedRecordMap } from 'notion-types'
 import { getPageTitle } from 'notion-utils'
-import { NotionRenderer } from 'react-notion-x'
 
 export const NotionPage = ({
   recordMap,
