+++ packages/react-notion-x/src/types.ts
@@ -27,6 +27,25 @@ export interface NotionComponents {
   Code: any
   Equation: any
   Callout?: any
+  TextHeader?: any
+  Divider?: any
+  TextBlock?: any
+  Drive?: any
+  List?: any
+  ColumnList?: any
+  Column?: any
+  Quote?: any
+  Bookmark?: any
+  Toggle?: any
+  TableOfContents?: any
+  ToDo?: any
+  Alias?: any
+  Table?: any
+  TableRow?: any
+  Page?: any
+  EOI?: any
+  Audio?: any
+  File?: any
 
   // collection
   Collection: any
